feat: fix Sonar findings (#129)
* fix(cpp:S128): resolve SonarQube blocker issue

* fix(cpp:S1709): resolve SonarQube issues

* fix(cpp:S1709): resolve SonarQube issues

* fix(cpp:S1699): don't call virtual functions from ctors

* fix: CodeQL finding cpp/comparison-with-wider-type

* fix(cpp:S5416): use "using" for type aliasing

* fix(cpp:S1905): remove redundant casts

* fix: add missing header

* fix: remove unused variable

* Processed review comments

+++ services/network/Http.cpp
@@ -113,7 +113,7 @@ namespace services
     void HttpRequestFormatter::AddContentLength(std::size_t size)
     {
         infra::StringOutputStream contentLengthStream(contentLength);
-        contentLengthStream << static_cast<uint64_t>(size);
+        contentLengthStream << size;
         contentLengthHeader.Emplace("Content-Length", contentLength);
     }
